I am stucked.
Return true if the string in the first element of the array contains all of the letters of the string in the second element of the array.
For example, ["hello", "Hello"], should return true because all of the letters in the second string are present in the first, ignoring case.
This code does not work on several object. For instance Mary-Army , Mary-Aarmy etc.